+++ wxc_sdk/as_api.py
@@ -73,9 +73,9 @@ __all__ = ['AsAccessCodesApi', 'AsAdminAuditEventsApi', 'AsAgentCallerIdApi', 'A
            'AsPrivateNetworkConnectApi', 'AsPushToTalkApi', 'AsRebuildPhonesJobsApi', 'AsReceptionistApi',
            'AsReceptionistContactsDirectoryApi', 'AsRecordingsApi', 'AsReportsApi', 'AsRestSession', 'AsRoomTabsApi',
            'AsRoomsApi', 'AsRouteGroupApi', 'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2UsersApi', 'AsScheduleApi',
-           'AsScimApiChild', 'AsScimV2Api', 'AsSelectiveAcceptApi', 'AsSelectiveRejectApi', 'AsSequentialRingApi',
-           'AsSimRingApi', 'AsStatusAPI', 'AsTeamMembershipsApi', 'AsTeamsApi', 'AsTelephonyApi',
-           'AsTelephonyDevicesApi', 'AsTelephonyLocationApi', 'AsTransferNumbersApi', 'AsTrunkApi',
+           'AsScimApiChild', 'AsScimV2Api', 'AsSelectiveAcceptApi', 'AsSelectiveForwardApi', 'AsSelectiveRejectApi',
+           'AsSequentialRingApi', 'AsSimRingApi', 'AsStatusAPI', 'AsTeamMembershipsApi', 'AsTeamsApi',
+           'AsTelephonyApi', 'AsTelephonyDevicesApi', 'AsTelephonyLocationApi', 'AsTransferNumbersApi', 'AsTrunkApi',
            'AsVirtualLinesApi', 'AsVoiceMessagingApi', 'AsVoicePortalApi', 'AsVoicemailApi', 'AsVoicemailGroupsApi',
            'AsVoicemailRulesApi', 'AsWebexSimpleApi', 'AsWebhookApi', 'AsWorkspaceDevicesApi',
            'AsWorkspaceLocationApi', 'AsWorkspaceLocationFloorApi', 'AsWorkspaceNumbersApi',
@@ -5901,6 +5901,7 @@ class AsPersonSettingsApiChild(AsApiChild, base=''):
             ('workspaces', 'priorityAlert'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'pushToTalk'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'selectiveAccept'): ('telephony/config/workspaces', '/'),
+            ('workspaces', 'selectiveForward'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'selectiveReject'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'sequentialRing'): ('telephony/config/workspaces', '/'),
             ('workspaces', 'simultaneousRing'): ('telephony/config/workspaces', '/'),
@@ -20600,6 +20601,203 @@ class AsSelectiveAcceptApi(AsPersonSettingsApiChild):
         await super().put(url, params=params, json=body)
 
 
+class AsSelectiveForwardApi(AsPersonSettingsApiChild):
+    """
+    API for selective forward settings
+
+    For now only used for workspaces
+    """
+
+    feature = 'selectiveForward'
+
+    async def read_criteria(self, entity_id: str, id: str,
+                      org_id: str = None) -> SelectiveForwardCriteria:
+        """
+        Retrieve Selective Forward Criteria for a Workspace
+
+        Retrieve Selective Forward Criteria Settings for a Workspace.
+
+        With the Selective Forward feature, you can forward calls at specific times from specific callers. This setting
+        takes precedence over call forwarding.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
+
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
+        :param id: Unique identifier for the criteria.
+        :type id: str
+        :param org_id: ID of the organization within which the entity resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: :class:`SelectiveCriteria`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.f_ep(entity_id, f'criteria/{id}')
+        data = await super().get(url, params=params)
+        r = SelectiveForwardCriteria.model_validate(data)
+        return r
+
+    async def configure_criteria(self, entity_id: str, id: str, settings: SelectiveForwardCriteria,
+                           org_id: str = None):
+        """
+        Modify Selective Forward Criteria for a Workspace
+
+        Modify Selective Forward Call Criteria Settings for a Workspace.
+
+        With the Selective Forward feature, you can forward calls at specific times from specific callers. This setting
+        takes precedence over call forwarding.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
+        :param id: Unique identifier for the criteria.
+        :type id: str
+        :param settings: new settings to be applied.
+        :type settings: SelectiveForwardCriteria
+        :param org_id: ID of the organization within which the entity resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = settings.update()
+        url = self.f_ep(entity_id, f'criteria/{id}')
+        await super().put(url, params=params, json=body)
+
+    async def delete_criteria(self, entity_id: str, id: str, org_id: str = None):
+        """
+        Delete Selective Forward Criteria for a Workspace
+
+        Delete Selective Forward Call criteria Settings for a workspace.
+
+        With the Selective Forward feature, you can forward calls at specific times from specific callers. This setting
+        takes precedence over call forwarding.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
+        :param id: Unique identifier for the criteria.
+        :type id: str
+        :param org_id: ID of the organization within which the entity resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.f_ep(entity_id, f'criteria/{id}')
+        await super().delete(url, params=params)
+
+    async def create_criteria(self, entity_id: str, settings: SelectiveForwardCriteria,
+                        org_id: str = None) -> str:
+        """
+        Create Selective Forward Criteria for a Workspace
+
+        Create Selective Forward Call Criteria Settings for a Workspace.
+
+        With the Selective Forward feature, you can forward calls at specific times from specific callers. This setting
+        takes precedence over call forwarding.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
+
+        This API requires a full, user or location administrator auth token with the `spark-admin:workspaces_write`
+        scope or a user auth token with a scope of `spark:workspaces_write` to update workspace settings.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
+        :param settings: new settings to be applied.
+        :type settings: SelectiveCriteria
+        :param org_id: ID of the organization within which the entity resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: str
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = settings.update()
+
+        url = self.f_ep(entity_id, f'criteria')
+        data = await super().post(url, params=params, json=body)
+        r = data['id']
+        return r
+
+    async def read(self, entity_id: str,
+             org_id: str = None) -> SelectiveForward:
+        """
+        Retrieve Selective Forward Settings for a Workspace
+
+        Retrieve Selective Forward Call Settings for a Workspace.
+
+        With the Selective Forward feature, you can forward calls at specific times from specific callers. This setting
+        takes precedence over call forwarding.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
+        :param org_id: ID of the organization within which the entity resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: :class:`SelectiveForward`
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.f_ep(entity_id)
+        data = await super().get(url, params=params)
+        r = SelectiveForward.model_validate(data)
+        return r
+
+    async def configure(self, entity_id: str, settings: SelectiveForward,
+                  org_id: str = None):
+        """
+        Modify Selective Forward Settings for a Workspace
+
+        Modify Selective Forward Call Settings for a Workspace.
+
+        With the Selective Forward feature, you can forward calls at specific times from specific callers. This setting
+        takes precedence over call forwarding.
+        Schedules can also be set up for this feature during certain times of the day or days of the week.
+
+        **NOTE**: This API is only available for professional licensed workspaces.
+
+        :param entity_id: Unique identifier for the entity.
+        :type entity_id: str
+        :param settings: Settings for new criteria
+        :type settings: SelectiveForward
+        :param org_id: ID of the organization within which the entity resides. Only admin users of another
+            organization (such as partners) may use this parameter as the default is the same organization as the
+            token used to access API.
+        :type org_id: str
+        :rtype: None
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = settings.update()
+        url = self.f_ep(entity_id)
+        await super().put(url, params=params, json=body)
+
+
 class AsSelectiveRejectApi(AsPersonSettingsApiChild):
     """
     API for selective reject settings
@@ -21320,6 +21518,7 @@ class AsWorkspaceSettingsApi(AsApiChild, base='workspaces'):
     privacy: AsPrivacyApi
     push_to_talk: AsPushToTalkApi
     selective_accept: AsSelectiveAcceptApi
+    selective_forward: AsSelectiveForwardApi
     selective_reject: AsSelectiveRejectApi
     sequential_ring: AsSequentialRingApi
     sim_ring: AsSimRingApi
@@ -21347,6 +21546,7 @@ class AsWorkspaceSettingsApi(AsApiChild, base='workspaces'):
         self.privacy = AsPrivacyApi(session=session, selector=ApiSelector.workspace)
         self.push_to_talk = AsPushToTalkApi(session=session, selector=ApiSelector.workspace)
         self.selective_accept = AsSelectiveAcceptApi(session=session, selector=ApiSelector.workspace)
+        self.selective_forward = AsSelectiveForwardApi(session=session, selector=ApiSelector.workspace)
         self.selective_reject = AsSelectiveRejectApi(session=session, selector=ApiSelector.workspace)
         self.sequential_ring = AsSequentialRingApi(session=session, selector=ApiSelector.workspace)
         self.sim_ring = AsSimRingApi(session=session, selector=ApiSelector.workspace)