+++ wxc_sdk/workspace_settings/__init__.py
@@ -26,6 +26,7 @@ from ..person_settings.priority_alert import PriorityAlertApi
 from ..person_settings.privacy import PrivacyApi
 from ..person_settings.push_to_talk import PushToTalkApi
 from ..person_settings.selective_accept import SelectiveAcceptApi
+from ..person_settings.selective_forward import SelectiveForwardApi
 from ..person_settings.selective_reject import SelectiveRejectApi
 from ..person_settings.sequential_ring import SequentialRingApi
 from ..person_settings.sim_ring import SimRingApi
@@ -64,6 +65,7 @@ class WorkspaceSettingsApi(ApiChild, base='workspaces'):
     privacy: PrivacyApi
     push_to_talk: PushToTalkApi
     selective_accept: SelectiveAcceptApi
+    selective_forward: SelectiveForwardApi
     selective_reject: SelectiveRejectApi
     sequential_ring: SequentialRingApi
     sim_ring: SimRingApi
@@ -91,6 +93,7 @@ class WorkspaceSettingsApi(ApiChild, base='workspaces'):
         self.privacy = PrivacyApi(session=session, selector=ApiSelector.workspace)
         self.push_to_talk = PushToTalkApi(session=session, selector=ApiSelector.workspace)
         self.selective_accept = SelectiveAcceptApi(session=session, selector=ApiSelector.workspace)
+        self.selective_forward = SelectiveForwardApi(session=session, selector=ApiSelector.workspace)
         self.selective_reject = SelectiveRejectApi(session=session, selector=ApiSelector.workspace)
         self.sequential_ring = SequentialRingApi(session=session, selector=ApiSelector.workspace)
         self.sim_ring = SimRingApi(session=session, selector=ApiSelector.workspace)