+++ docs/changes.rst
@@ -35,6 +35,7 @@ Release history
 - feat: selective reject settings API for workspaces: :attr:`api.workspace_settings.selective_reject <wxc_sdk.workspace_settings.WorkspaceSettingsApi.selective_reject>`
 - feat: selective accept settings API for workspaces: :attr:`api.workspace_settings.selective_accept <wxc_sdk.workspace_settings.WorkspaceSettingsApi.selective_accept>`
 - feat: priority alert settings API for workspaces: :attr:`api.workspace_settings.priority_alert <wxc_sdk.workspace_settings.WorkspaceSettingsApi.priority_alert>`
+- feat: selective forward settings API for workspaces: :attr:`api.workspace_settings.selective_forward <wxc_sdk.workspace_settings.WorkspaceSettingsApi.selective_forward>`
 
 1.19.0
 ------